requirements.txt
boto3
pandas
pyarrow
pyyaml
streamlit
plotly
fastapi
uvicorn[standard]
requests
psutil
fsspec
s3fs
